adiciona o esqueleto do projeto no main. Imports do framebuffer, color e algoritmo de bresenham
