+++ main.js
@@ -1,7 +1,32 @@
+// IMPORTS
+import clickableGrid from './classes/utils';
+let Color = require ('./classes/color');
+let FrameBuffer = require ('./classes/framebuffer');
+let Algorithms = require('./algorithms');
+
+// CONFIGS
+const WIDTH = 35;
+const HEIGHT = 50;
+
+// Control variable
 let countClicks = 0;
+
+// Grid Related
+let frameBuffer = new FrameBuffer(WIDTH,HEIGHT);
 let startCoordinates = {};
 let endCoordinates = {};
-var grid = clickableGrid(35,50,function(el,x,y){
+
+function paintPoints(){
+    for (let y=0; y<frameBuffer.height; y++){
+        for (let x=0; x<frameBuffer.width; x++){
+            let pixel = document.getElementById(x + "-" + y);
+            pixel.style.backgroundColor = frameBuffer.getPixel(x,y).color.getRGB();
+
+        }
+    }
+}
+
+var grid = clickableGrid(WIDTH,HEIGHT,function(el,x,y){
     console.log("You clicked on element:",el);
     console.log("You clicked on item #:",x,y);
 
@@ -13,7 +38,13 @@ var grid = clickableGrid(35,50,function(el,x,y){
         endCoordinates.x = x;
         endCoordinates.y = y;
         // paint
-
+        // TODO: check which algorithm should be used
+        let pixelsToPaint = Algorithms.bresenham(startCoordinates,endCoordinates);
+        console.log("Finished Bresenham");
+        for (let i=0; i < pixelsToPaint.length -1 ; i+=2){
+            frameBuffer.getPixel(pixelsToPaint[i],pixelsToPaint[i+1]).color = new Color(200,0,0);
+        }
+        paintCanvas();
         // reset
         countClicks = 0;
         startCoordinates = {};
@@ -31,26 +62,7 @@ var grid = clickableGrid(35,50,function(el,x,y){
 
 document.body.appendChild(grid);
 
-function clickableGrid( rows, cols, callback ){
-    var grid = document.createElement('table');
-    var x,y;
-    grid.className = 'grid';
-    for (var r=0;r<rows;++r){
-        var tr = grid.appendChild(document.createElement('tr'));
-        for (var c=0;c<cols;++c){
-            var cell = tr.appendChild(document.createElement('td'));
-            x = c;
-            y = r;
-            cell.id = x + "-" + y;
-            cell.addEventListener('click',(function(el,x,y){
-                return function(){
-                    callback(el,x,y);
-                }
-            })(cell,x,y),false);
-        }
-    }
-    return grid;
-}
+
